+++ packages/framer-motion/src/animation/animators/utils/can-animate.ts
@@ -1,5 +1,5 @@
 import { ResolvedKeyframes } from "../../../render/utils/KeyframesResolver"
-import { warning } from "../../../utils/errors"
+import { warning } from "motion-utils"
 import { isGenerator } from "../../generators/utils/is-generator"
 import { AnimationGeneratorType } from "../../types"
 import { isAnimatable } from "../../utils/is-animatable"