+++ packages/framer-motion/src/events/event-info.ts
@@ -6,14 +6,11 @@ export type EventListenerWithPointInfo = (
     info: EventInfo
 ) => void
 
-export function extractEventInfo(
-    event: PointerEvent,
-    pointType: "page" | "client" = "page"
-): EventInfo {
+export function extractEventInfo(event: PointerEvent): EventInfo {
     return {
         point: {
-            x: event[`${pointType}X`],
-            y: event[`${pointType}Y`],
+            x: event.pageX,
+            y: event.pageY,
         },
     }
 }