+++ packages/framer-motion/src/frameloop/frame.ts
@@ -1,4 +1,4 @@
-import { noop } from "../utils/noop"
+import { noop } from "motion-utils"
 import { createRenderBatcher } from "./batcher"
 
 export const {