+++ packages/framer-motion/src/render/dom/resize/index.ts
@@ -1,4 +1,4 @@
-import { ElementOrSelector } from "../../../animation/types"
+import { ElementOrSelector } from "motion-dom"
 import { resizeElement } from "./handle-element"
 import { resizeWindow } from "./handle-window"
 import { ResizeHandler } from "./types"