+++ packages/motion-utils/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./errors"
+export * from "./noop"