+++ CHANGELOG.md
@@ -4,11 +4,24 @@ Motion adheres to [Semantic Versioning](http://semver.org/).
 
 Undocumented APIs should be considered internal and may change without warning.
 
-## [11.11.17] 2024-10-26
+## [11.13.1] 2024-12-03
 
 ### Fixed
 
 - SVG elements (like `motion.text`) now update when given a `MotionValue` as children, matching HTML element behavior.
+
+## [11.13.0] 2024-12-03
+
+### Added
+
+-   Vanilla `hover` gesture in Motion+ early access.
+
+## [11.12.0] 2024-11-27
+
+### Added
+
+-   New `visualDuration` option for `spring` animations.
+-   New `spring(visualDuration, bounce)` syntax.
 
 ## [11.11.16] 2024-11-14